Cluster-GKE/requirements.txt
flask==3.0.2
gunicorn==20.1.0
google-cloud-pubsub==2.19.7
google-cloud-storage==2.14.0
